requirements.txt
streamlit
pillow
kraken
